requirements.txt
cycler==0.9.0
decorator==4.0.4
Django==1.9
django-cleanup==0.4.1
django-filter==0.11.0
djangorestframework==3.3.1
docopt==0.6.2
funcsigs==0.4
gdbn==0.1
gnumpy==0.2
Jinja2==2.8
joblib==0.9.2
-e git+https://github.com/Lasagne/Lasagne.git@b6f1250eb95b6b036cd348a37dbfc4dbf84d8a67#egg=Lasagne-dev
mahotas==1.4.0
Markdown==2.6.5
MarkupSafe==0.23
matplotlib==1.4.3
mock==1.3.0
networkx==1.10
nolearn==0.6a0.dev0
nose==1.3.7
numpy==1.10.1
pbr==1.8.1
Pillow==3.0.0
pyparsing==2.0.5
python-dateutil==2.4.2
pytz==2015.7
scikit-image==0.11.3
scikit-learn==0.16.1
scipy==0.16.1
six==1.10.0
tabulate==0.7.5
-e git+https://github.com/Theano/Theano.git@15c90dd36c95ea2989a2c0085399147d15a6a477#egg=Theano-dev
wheel==0.24.0
